♻️(ats-presentation) improve ats features folder structure
