+++ src/web/presentation/frontend/src/features/ats/constants/etape-recrutement.ts
@@ -1,4 +1,4 @@
-import type { EtapeRecrutement } from '../api/recrutement'
+import type { EtapeRecrutement } from '../parametres/api'
 
 type Categorie = EtapeRecrutement['categorie']
 type BadgeType = 'info' | 'success' | 'error'
