+++ src/web/presentation/frontend/src/features/ats/parametres/api.ts
@@ -4,14 +4,6 @@ import { api } from '@/utils/api'
 export type EtapeRecrutement = components['schemas']['EtapeRecrutement']
 export type UpdateEtapeRecrutement = components['schemas']['UpdateEtapeRecrutement']
 
-export type PaginatedRecrutements = components['schemas']['PaginatedRecrutementsResponse']
-
-export interface RecrutementsQuery {
-  filtre?: 'actifs' | 'archives'
-  page?: number
-  size?: number
-}
-
 export async function getEtapesRecrutement(organismeUuid: string): Promise<EtapeRecrutement[]> {
   const { data } = await api.GET('/recruteur/organisme/{organisme_uuid}/parametres/etapes', {
     params: { path: { organisme_uuid: organismeUuid } },
@@ -33,16 +25,6 @@ export async function updateEtapesRecrutement(
 export async function initEtapesRecrutement(organismeUuid: string): Promise<EtapeRecrutement[]> {
   const { data } = await api.POST('/recruteur/organisme/{organisme_uuid}/parametres/etapes/init', {
     params: { path: { organisme_uuid: organismeUuid } },
-  })
-  return data!
-}
-
-export async function getRecrutements(
-  organismeUuid: string,
-  query: RecrutementsQuery = {},
-): Promise<PaginatedRecrutements> {
-  const { data } = await api.GET('/recruteur/organisme/{organisme_uuid}/recrutements', {
-    params: { path: { organisme_uuid: organismeUuid }, query },
   })
   return data!
 }
