+++ src/web/presentation/frontend/src/features/ats/recrutements/filters.test.ts
@@ -1,4 +1,4 @@
-import type { RecrutementBase } from '../types'
+import type { RecrutementBase } from './types'
 import { describe, expect, it } from 'vitest'
 import {
   emptyRecrutementsFilters,
